+++ src/main/java/com/ynyes/lyz/controller/management/TdManagerReturnNoteController.java
@@ -38,7 +38,7 @@ public class TdManagerReturnNoteController {
 				            String __VIEWSTATE,
 				            Long[] listId,
 				            Integer[] listChkId,
-				            Long[] listSortId,
+				            Double[] listSortId,
 				            ModelMap map,
 				            HttpServletRequest req){
 			String username = (String) req.getSession().getAttribute("manager");
@@ -189,7 +189,7 @@ public class TdManagerReturnNoteController {
 	        }
 	    }
 		
-		 private void btnSave(String type, Long[] ids, Long[] sortIds)
+		 private void btnSave(String type, Long[] ids, Double[] sortIds)
 		    {
 		        if (null == type || type.isEmpty())
 		        {
