+++ src/main/java/com/ynyes/lyz/controller/management/TdManagerSettingController.java
@@ -124,7 +124,7 @@ public class TdManagerSettingController {
                         String __VIEWSTATE,
                         Long[] listId,
                         Integer[] listChkId,
-                        Long[] listSortId,
+                        Double[] listSortId,
                         ModelMap map,
                         HttpServletRequest req){
         String username = (String) req.getSession().getAttribute("manager");
@@ -946,7 +946,7 @@ public class TdManagerSettingController {
         }
     }    
     
-    private void btnSave(Long[] ids, Long[] sortIds)
+    private void btnSave(Long[] ids, Double[] sortIds)
     {
         if (null == ids || null == sortIds
                 || ids.length < 1 || sortIds.length < 1)
